requirements.txt
Flask==1.1.1
google-cloud-datastore==1.10.0
flask-restplus==0.13.0
qrcode==6.1
